created separate bnode class with method headers. adjusted sizing calculations of bnode

+++ params.h
@@ -1,8 +1,10 @@
 #ifndef _PARAMS_H_
 
+#include "stddef.h"
+
 /*
  * Number of bytes in a page
  */
-static const unsigned int PAGE_SIZE_IN_BYTES = 4096;
+static const size_t PAGE_SIZE_IN_BYTES = 4096;
 
 #endif
